finaly got it debugged :)
GamePage.jsx is showing player's names with their respective hands below it along with the game score being rendering and the message to select staring players. Fixed a lot of logic for when/how to access the in game Hand dictionary that is updated throughout the game and the database that only stores the initial hand (for now just for reference on the GameStats page, eventually will be used to replay the same game withe the same starting hands.) Worked on separation of concerns on which layers should be adding to the database vs which should be updating the in game Hand. fixed database to store JSON string